+++ main_test_util.go
@@ -17,6 +17,7 @@ const oneTime = "0001-01-01T01:00:00Z"
 const envName01 = "envName01"
 const envName02 = "envName02"
 const envVarName01 = "envVarName01"
+const envVarValue01 = "envVarValue01"
 const envRefName01 = "envRefName01"
 
 func createTempDB(t *testing.T) string {